reestructuracion de componentes/carpetas
reestructure los componentes en carpetas para tener un mejor orden y poder diferenciar bien las funciones
