+++ apps/estacionamientos/forms/reservas.py
@@ -1,15 +1,6 @@
 from django import forms
 from django.utils import timezone
-from .models import Estacionamiento, Reserva
-
-class EstacionamientoForm(forms.ModelForm):
-    class Meta:
-        model = Estacionamiento
-        fields = ["estado", "tipo"]
-
-class EstacionamientosMasivoForm(forms.Form):
-    cantidad = forms.IntegerField(min_value=1, initial=1)
-    tipo = forms.ChoiceField(choices=Estacionamiento.TIPO_CHOICES)
+from ..models.reserva import Reserva  
 
 class ReservaCrearForm(forms.Form):
     patente = forms.CharField(max_length=10)
@@ -19,10 +10,10 @@ class ReservaCrearForm(forms.Form):
 
     def clean(self):
         data = super().clean()
-        if "fecha_inicio" in data:
-            if data["fecha_inicio"].date() == timezone.now().date():
+        if "fecha_inicio" in data and data["fecha_inicio"] is not None:
+            now = timezone.now()
+            if data["fecha_inicio"].date() == now.date():
                 raise forms.ValidationError("No puedes reservar para el mismo día.")
-            if data["fecha_inicio"] < timezone.now():
+            if data["fecha_inicio"] < now:
                 raise forms.ValidationError("La fecha no puede ser pasada.")
         return data
-
