+++ apps/estacionamientos/urls.py
@@ -1,24 +1,32 @@
 from django.urls import path
-from . import views
+
+from .views.home import home, marcarSalida, marcarSalidaPatente
+from .views.estacionamientos import (
+    listarEstacionamiento, crearEstacionamiento, crearEstacionamientosMasivo,
+    editarEstacionamiento, eliminarEstacionamiento, eliminarTodosEstacionamientos,
+)
+from .views.reservas import listarReserva, crearReserva, terminarReserva
+from .views.historial import listarHistorial
 
 urlpatterns = [
     # Home 
-    path("", views.home, name="home"),
-    path("salida/<int:id>/", views.marcarSalida, name="marcarSalida"),
-    path("salida_por_patente/", views.marcarSalidaPatente, name="marcarSalidaPatente"),
+    path("", home, name="home"),
+    path("salida/<int:id>/", marcarSalida, name="marcarSalida"),
+    path("salida_por_patente/", marcarSalidaPatente, name="marcarSalidaPatente"),
 
     # CRUD Estacionamientos
-    path("estacionamientos/", views.listarEstacionamiento, name="listarEstacionamiento"),
-    path("estacionamientos/crear/", views.crearEstacionamiento, name="crearEstacionamiento"),
-    path("estacionamientos/crear_masivo/", views.crearEstacionamientosMasivo, name="crearEstacionamientosMasivo"),
-    path("estacionamientos/<int:id>/editar/", views.editarEstacionamiento, name="editarEstacionamiento"),
-    path("estacionamientos/<int:id>/eliminar/", views.eliminarEstacionamiento, name="eliminarEstacionamiento"),
-    path("estacionamientos/eliminar_todos/", views.eliminarTodosEstacionamientos, name="eliminarTodosEstacionamientos"),
+    path("estacionamientos/", listarEstacionamiento, name="listarEstacionamiento"),
+    path("estacionamientos/crear/", crearEstacionamiento, name="crearEstacionamiento"),
+    path("estacionamientos/crear_masivo/", crearEstacionamientosMasivo, name="crearEstacionamientosMasivo"),
+    path("estacionamientos/<int:id>/editar/", editarEstacionamiento, name="editarEstacionamiento"),
+    path("estacionamientos/<int:id>/eliminar/", eliminarEstacionamiento, name="eliminarEstacionamiento"),
+    path("estacionamientos/eliminar_todos/", eliminarTodosEstacionamientos, name="eliminarTodosEstacionamientos"),
 
     # Reservas
-    path("reservas/", views.listarReserva, name="listarReserva"),
-    path("reservas/crear/", views.crearReserva, name="crearReserva"),
-    path("reservas/<int:id>/terminar/", views.terminarReserva, name="terminarReserva"),
+    path("reservas/", listarReserva, name="listarReserva"),
+    path("reservas/crear/", crearReserva, name="crearReserva"),
+    path("reservas/<int:id>/terminar/", terminarReserva, name="terminarReserva"),
 
-    path("historial/", views.listarHistorial, name="listarHistorial"),
+    # Historial
+    path("historial/", listarHistorial, name="listarHistorial"),
 ]
